requirements.txt
pydantic~=1.10.12
pydantic-argparse~=0.8.0
rich~=13.5.3
ruamel.yaml~=0.17.32
requests~=2.31.0
jsonstream~=0.0.1
python_dateutil~=2.8.2
